refactor: move slots queries and mutations to  slot graphql module
