+++ src/graphql/schema/day/mutations/index.ts
@@ -1,3 +1,1 @@
 export * from './createDay'
-export * from './createSlot'
-export * from './makeReservation'
