+++ src/graphql/schema/day/queries/index.ts
@@ -1,3 +1,2 @@
 
 export * from './get-day'
-export * from './get-slot'
