+++ src/graphql/schema/slot/mutations/create-slot.ts
