+++ src/graphql/schema/slot/mutations/index.ts
@@ -1,0 +1,2 @@
+export * from './create-slot'
+export * from './make-reservation'
