+++ src/graphql/schema/slot/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './get-slot'
