+++ src/graphql/schema/ticket/types/index.ts
@@ -1,0 +1,1 @@
+export * from './ticket'
